Resolucion de los ejercicios Floating point exception y segmentation fault guardados dentro de Respuestas.md

+++ debug/fpe/test_fpe3.c
@@ -1,5 +1,6 @@
 #include <stdlib.h>
 #include <stdio.h>
+#include <math.h>
 #ifdef TRAPFPE
 #include "fpe_x87_sse.h"
 #endif
@@ -16,7 +17,7 @@ int main(int argc, char *argv[])
   scanf("%f",&a);
   scanf("%f",&b);
 
-  tmp = sqrtf(a / b);
+  tmp = sqrt(a / b);
   
   if(tmp > 2.)
     {
